データを保存する機能実装
