--- app/helpers/transactions_helper.rb
@@ -1,2 +1,0 @@
-module TransactionsHelper
-end
