+++ config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
   
   resources :users 
   resources :items do
-    resources :transactions, only: [:index, :new, :create]
+    resources :buyers, only: [:index, :create]
   end
 
 end
