+++ spec/requests/buyers_request_spec.rb
@@ -2,18 +2,4 @@ require 'rails_helper'
 
 RSpec.describe "Buyers", type: :request do
 
-  describe "GET /index" do
-    it "returns http success" do
-      get "/buyers/index"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /done" do
-    it "returns http success" do
-      get "/buyers/done"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
 end
