--- spec/requests/transactions_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Transactions", type: :request do
-
-end
